+++ centreon/packages/js-config/cypress/e2e/plugins.ts
@@ -3,6 +3,31 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 /* eslint-disable no-param-reassign */
 
+import fs from 'fs';
+import path from 'path';
+
+const handleTestResults = async (results: CypressCommandLine.RunResult) => {
+  const testRetries: { [key: string]: boolean } = {};
+  if (results && results.tests) {
+    results.tests.forEach((test) => {
+      if (test.attempts && test.attempts.length > 1) {
+        const testTitle = test.title.join(' > ');
+        testRetries[testTitle] = true;
+      }
+    });
+  }
+
+  console.log('Test retries:', testRetries);
+  if (Object.keys(testRetries).length > 0) {
+    const resultFilePath = path.join(
+      __dirname,
+      '../../../../tests/e2e/results',
+      'hasRetries.json'
+    );
+    fs.writeFileSync(resultFilePath, JSON.stringify(testRetries, null, 2));
+  }
+};
+
 export default (
   on: Cypress.PluginEvents,
   config: Cypress.PluginConfigOptions
@@ -30,6 +55,10 @@ export default (
     }
 
     return launchOptions;
+  });
+
+  on('after:spec', async (spec, results) => {
+    await handleTestResults(results);
   });
 
   return config;
